+++ ikea-zigbee-drivers/Ikea_E2013.groovy
@@ -9,7 +9,7 @@ import groovy.transform.Field
 import com.hubitat.zigbee.DataType
 
 @Field static final String DRIVER_NAME = 'IKEA Parasoll Door/Window Sensor (E2013)'
-@Field static final String DRIVER_VERSION = '5.1.0'
+@Field static final String DRIVER_VERSION = '5.2.0'
 
 // Fields for capability.IAS
 import hubitat.zigbee.clusters.iaszone.ZoneStatus
@@ -57,7 +57,7 @@ metadata {
             name:'helpInfo', type:'hidden',
             title:'''
             <div style="min-height:55px; background:transparent url('https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/img/Ikea_E2013.webp') no-repeat left center;background-size:auto 55px;padding-left:60px">
-                IKEA Parasoll Door/Window Sensor (E2013) <small>v5.1.0</small><br>
+                IKEA Parasoll Door/Window Sensor (E2013) <small>v5.2.0</small><br>
                 <small><div>
                 • <a href="https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/#parasoll-doorwindow-sensor-e2013" target="_blank">device details</a><br>
                 • <a href="https://community.hubitat.com/t/release-ikea-zigbee-drivers/123853" target="_blank">community page</a><br>